requirements.txt
arrow==1.1.0
asgiref @ file:///tmp/build/80754af9/asgiref_1617740417847/work
binaryornot==0.4.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
cookiecutter==1.7.3
crispy-tailwind==0.5.0
cryptography==35.0.0
defusedxml==0.7.1
Django~=2.2.5
django-allauth~=0.49.0
django-braces==1.14.0
django-crispy-forms==1.12.0
django-environ==0.7.0
django-tailwind==2.0.1
fast-enum==1.3.0
idna==3.3
Jinja2==3.0.1
jinja2-time==0.2.0
Logbook==1.5.3
MarkupSafe==2.0.1
misaka==2.1.1
oauthlib==3.1.1
Pillow==8.2.0
poyo==0.5.0
psycopg2 @ file:///C:/ci/psycopg2_1612298766762/work
pycparser==2.20
PyJWT==2.2.0
python-dateutil==2.8.1
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
single-version==1.5.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
Unidecode==1.2.0
urllib3==1.26.7
wincertstore==0.2

turtle~=0.0.1
faker~=13.3.2
django-filter~=21.1
pathlib~=1.0.1